Zen-Ship Application/requirements.txt
tensorflow
torchaudio
Pillow
matplotlib==2.1.1
tqdm
librosa
SoundFile
kivy
numpy
